[refactor pt 2] Move actions popover keyboard logic to `column_actions` file
- goal is to move all logic related to the actions popver in one file
