+++ packages/eui/src/components/datagrid/body/header/column_actions.test.tsx
@@ -6,7 +6,8 @@
  * Side Public License, v 1.
  */
 
-import { ReactElement } from 'react';
+import React, { ReactElement } from 'react';
+import { render, renderHook } from '../../../../test/rtl';
 
 import { EuiListGroupItemProps } from '../../../list_group';
 import { schemaDetectors } from '../../utils/data_grid_schema';
@@ -15,7 +16,138 @@ import {
   getColumnActions,
   isColumnActionEnabled,
   getColumnActionConfig,
+  usePopoverArrowNavigation,
 } from './column_actions';
+
+describe('usePopoverArrowNavigation', () => {
+  const MockPanel = () => (
+    <div tabIndex={-1}>
+      <button data-test-subj="first">First action</button>
+      <button data-test-subj="second">Second action</button>
+      <button data-test-subj="last">Last action</button>
+    </div>
+  );
+
+  const {
+    panelRef,
+    panelProps: { onKeyDown },
+  } = renderHook(usePopoverArrowNavigation).result.current;
+
+  let mockPanel: HTMLElement;
+
+  const preventDefault = jest.fn();
+  const keyDownEvent = { preventDefault } as unknown as React.KeyboardEvent;
+  beforeEach(() => jest.clearAllMocks());
+
+  describe('early returns', () => {
+    it('does nothing if the up/down arrow keys are not pressed', () => {
+      onKeyDown({ ...keyDownEvent, key: 'Tab' });
+      expect(preventDefault).not.toHaveBeenCalled();
+    });
+
+    it('does nothing if the popover contains no tabbable elements', () => {
+      const emptyDiv = document.createElement('div');
+      panelRef(emptyDiv);
+      onKeyDown({ ...keyDownEvent, key: 'ArrowDown' });
+      expect(preventDefault).not.toHaveBeenCalled();
+
+      panelRef(mockPanel); // Reset for other tests
+    });
+  });
+
+  describe('when the popover panel is focused (on initial open state)', () => {
+    beforeEach(() => {
+      const { container } = render(<MockPanel />);
+
+      mockPanel = container.firstElementChild as HTMLElement;
+      panelRef(mockPanel);
+
+      mockPanel.focus();
+    });
+    it('focuses the first action when the arrow down key is pressed', () => {
+      expect(mockPanel).toHaveFocus();
+      onKeyDown({ ...keyDownEvent, key: 'ArrowDown' });
+      expect(preventDefault).toHaveBeenCalled();
+      expect(document.activeElement?.getAttribute('data-test-subj')).toEqual(
+        'first'
+      );
+    });
+
+    it('focuses the last action when the arrow up key is pressed', () => {
+      onKeyDown({ ...keyDownEvent, key: 'ArrowUp' });
+      expect(preventDefault).toHaveBeenCalled();
+      expect(document.activeElement?.getAttribute('data-test-subj')).toEqual(
+        'last'
+      );
+    });
+  });
+
+  describe('when already focused on action buttons', () => {
+    describe('down arrow key', () => {
+      beforeEach(() => {
+        const { container } = render(<MockPanel />);
+
+        mockPanel = container.firstElementChild as HTMLElement;
+        panelRef(mockPanel);
+      });
+
+      it('moves focus to the the next action', () => {
+        (mockPanel.firstElementChild as HTMLButtonElement).focus();
+
+        onKeyDown({ ...keyDownEvent, key: 'ArrowDown' });
+        expect(document.activeElement?.getAttribute('data-test-subj')).toEqual(
+          'second'
+        );
+
+        onKeyDown({ ...keyDownEvent, key: 'ArrowDown' });
+        expect(document.activeElement?.getAttribute('data-test-subj')).toEqual(
+          'last'
+        );
+      });
+
+      it('loops focus back to the first action when pressing down on the last action', () => {
+        (mockPanel.lastElementChild as HTMLButtonElement).focus();
+
+        onKeyDown({ ...keyDownEvent, key: 'ArrowDown' });
+        expect(document.activeElement?.getAttribute('data-test-subj')).toEqual(
+          'first'
+        );
+      });
+    });
+
+    describe('up arrow key', () => {
+      beforeEach(() => {
+        const { container } = render(<MockPanel />);
+
+        mockPanel = container.firstElementChild as HTMLElement;
+        panelRef(mockPanel);
+      });
+
+      it('moves focus to the previous action', () => {
+        (mockPanel.lastElementChild as HTMLButtonElement).focus();
+
+        onKeyDown({ ...keyDownEvent, key: 'ArrowUp' });
+        expect(document.activeElement?.getAttribute('data-test-subj')).toEqual(
+          'second'
+        );
+
+        onKeyDown({ ...keyDownEvent, key: 'ArrowUp' });
+        expect(document.activeElement?.getAttribute('data-test-subj')).toEqual(
+          'first'
+        );
+      });
+
+      it('loops focus back to the last action when pressing up on the first action', () => {
+        (mockPanel.firstElementChild as HTMLButtonElement).focus();
+
+        onKeyDown({ ...keyDownEvent, key: 'ArrowUp' });
+        expect(document.activeElement?.getAttribute('data-test-subj')).toEqual(
+          'last'
+        );
+      });
+    });
+  });
+});
 
 describe('getColumnActions', () => {
   const setVisibleColumns = jest.fn();
