+++ packages/eui/src/components/datagrid/body/header/data_grid_header_cell.tsx
@@ -17,16 +17,17 @@ import React, {
   memo,
   KeyboardEventHandler,
 } from 'react';
-import { tabbable, FocusableElement } from 'tabbable';
+
 import { keys, useEuiMemoizedStyles } from '../../../../services';
-import { EuiI18n, useEuiI18n } from '../../../i18n';
+import { useEuiI18n } from '../../../i18n';
 import { EuiIcon } from '../../../icon';
 import { EuiListGroup } from '../../../list_group';
 import { EuiPopover } from '../../../popover';
 import { EuiButtonIcon } from '../../../button';
+
 import { DataGridFocusContext } from '../../utils/focus';
 import { EuiDataGridHeaderCellProps } from '../../data_grid_types';
-import { getColumnActions } from './column_actions';
+import { getColumnActions, usePopoverArrowNavigation } from './column_actions';
 import { useColumnSorting } from './column_sorting';
 import { EuiDataGridColumnResizer } from './data_grid_column_resizer';
 import { EuiDataGridHeaderCellWrapper } from './data_grid_header_cell_wrapper';
@@ -271,63 +272,3 @@ export const EuiDataGridHeaderCell: FunctionComponent<EuiDataGridHeaderCellProps
     }
   );
 EuiDataGridHeaderCell.displayName = 'EuiDataGridHeaderCell';
-
-/**
- * Add keyboard arrow navigation to the cell actions popover
- * to match the UX of the rest of EuiDataGrid
- */
-export const usePopoverArrowNavigation = () => {
-  const popoverPanelRef = useRef<HTMLElement | null>(null);
-  const actionsRef = useRef<FocusableElement[] | undefined>(undefined);
-  const panelRef = useCallback((ref: HTMLElement | null) => {
-    popoverPanelRef.current = ref;
-    actionsRef.current = ref ? tabbable(ref) : undefined;
-  }, []);
-
-  const onKeyDown = useCallback((e: React.KeyboardEvent) => {
-    if (e.key !== keys.ARROW_DOWN && e.key !== keys.ARROW_UP) return;
-    if (!actionsRef.current?.length) return;
-
-    e.preventDefault();
-
-    const initialState = document.activeElement === popoverPanelRef.current;
-    const currentIndex = !initialState
-      ? actionsRef.current.findIndex((el) => document.activeElement === el)
-      : -1;
-    const lastIndex = actionsRef.current.length - 1;
-
-    let indexToFocus: number;
-    if (initialState) {
-      if (e.key === keys.ARROW_DOWN) {
-        indexToFocus = 0;
-      } else if (e.key === keys.ARROW_UP) {
-        indexToFocus = lastIndex;
-      }
-    } else {
-      if (e.key === keys.ARROW_DOWN) {
-        indexToFocus = currentIndex + 1;
-        if (indexToFocus > lastIndex) {
-          indexToFocus = 0;
-        }
-      } else if (e.key === keys.ARROW_UP) {
-        indexToFocus = currentIndex - 1;
-        if (indexToFocus < 0) {
-          indexToFocus = lastIndex;
-        }
-      }
-    }
-
-    actionsRef.current[indexToFocus!].focus();
-  }, []);
-
-  return {
-    panelRef,
-    panelProps: { onKeyDown },
-    popoverScreenReaderText: (
-      <EuiI18n
-        token="euiDataGridHeaderCell.actionsPopoverScreenReaderText"
-        default="To navigate through the list of column actions, press the Tab or Up and Down arrow keys."
-      />
-    ),
-  };
-};
